+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- webapp: Fix PWA settings with downloadable sources
 - webapp: Fix custom prefix path with vite build
 
 ## [1.20.0] - 2025-08-14